accept input_file argument to site_generator$render
If a single file is passed to rmarkdown::render_site then this file is forwarded to the generator's render function as 'input_file'. This currently occurs when the IDE processes a "Knit" for a file within a website so we turn this into a single chapter preview.

Note that with this change knit: bookdown::preview_chapter will no longer be necessary to get single-file incremental preview in the IDE (this change should be in tomorrow's daily build of the IDE).