Document that \frontmatter goes into preamble.tex (#552)
This change brings the documentation in line with the implementation. It is important to have `\frontmatter` in `preamble.tex` and not in `before_body.tex`. Otherwise the title page would not be part of the front matter and page numbering would be reset *after* it. 

Motivated by this SO question: https://stackoverflow.com/questions/43711029/page-numbering-in-r-bookdown/