pdf_book: combine references and out.height When using the chunk option `out.height` the figure caption is translated differently.
Usually, all the content of the `includegraphics` command is separated into indiviual lines, like the path to the file, the label, the caption command, and the caption text itself. All occurrences of `\\ref()` have been replaced by `@ref()`.

But whenever the `out.height` option is present the whole `includegraphics` command will be returned in one line with the `\\@ref()` still intact. These will result to `\\ref{}` in the final document and cause LaTeX to throw an error.

This issue is resolved be adding an additional lines dealing with the `\\@ref()` instances arriving at the `pdf_book` function.