Change name of temporary files created.
Currently, during rending the code copies the backs up the R markdown file to a temporary file in the same directory with the a name like `bookdown11b6326822f9c`. This file should be cleaned up, but sometimes is not if the process is killed. If the file happens to not be cleaned up, it is hard to write a globbing expression for `.gitignore` and similar to match it that and wouldn't result in false positives; `bookdown*` is the best I can come up with. I propose

1. Use a `.bak` extension to make it easier to match with globbing.
2. Use the original filename as the base, since it is backing up a particular file. If the process is killed, then it is easier to restore the original file.

Alternatively, if the backup file never needs to be restored by the user, why not put it in the temporary directory?

Addresses #612