Add compatibility with new Pandoc 2.x way of creating self-contained document (#784)
allow to move multiline script node in html

As investigated in #739, Pandoc 2 modified the way it handles js scripts in self contained document. They are not always base64 encoded by default, and are rather included in `<script>` tags as-is. This means that for gitbook post processing, multiline blocks must be identified and moved.