Minor fix for bookdown::gitbook
In this Stackoverflow question, a minimal `bookdown::gitbook` document is created:  https://stackoverflow.com/q/62489019/2554330.  When viewing the result of the document in RStudio's previewer, I see this in the Javascript console:

```
Uncaught TypeError: Cannot read property 'getBoundingClientRect' of undefined
    at plugin-bookdown.js:152
    at Object.<anonymous> (plugin-bookdown.js:191)
    at Object.dispatch (jquery.min.js:3)
    at Object.r.handle (jquery.min.js:3)
    at Object.trigger (jquery.min.js:4)
    at Object.<anonymous> (jquery.min.js:4)
    at Function.each (jquery.min.js:2)
    at n.fn.init.each (jquery.min.js:2)
    at n.fn.init.trigger (jquery.min.js:4)
    at Object.start (app.min.js:1)
```

This happens because in that minimal document, `li[0]` is undefined.  The PR puts in a sanity check.  (There's a test for the length of `li` earlier in the code at line 105, but the change it makes in case of zero length, `li = chaps.first();`, doesn't solve the issue in the SO example.)