Enable pandoc `--file-scope` only when duplicate footnotes are detected
By default, the `--file-scope` argument is now only passed to pandoc when duplicate footnotes are detected in the manuscript. This is controlled by the `bookdown.render.file_scope` option, which now defaults to `"auto"`. Set the option to `TRUE` or `FALSE` to force a particular behavior irrespective of whether duplicates are detected.