fix #834: preserve input Rmd's origin path in opts
Remove input Rmd file's path in opts could bring inconsistency.
For example in `split_chapter()`, we compare `input_rmd` and `nms_chaps`
to determine if we could skip writing html. Then when we call
`serve_book()`, all the changed Rmd files under subdir will be ignored
due to the absence of `input_rmd`'s full path.
This is the root cause of #834.