:hammer: Adding color navigation in the sidebar
Colorful navigation between sections was added in the sidebar of the 'gitbook' format, using 'stylish' key in 'config'. It is possible to enter a valid RGB system color, in hexadecimal notation or an image (local or url), using in 'stylish' the 'cover_image' key. In the case of an image, the median of the color channels is considered. 'stylish' also has a key 'justify' that justifies all the text.