Accept custom template for `bs4_book()` (#1247)
to be the same as with `gitbook()`. 

Let's keep in mind that this is for advanced user only, because changing the template could break some features. It will be also hard to maintain if we make change to `bs4_book()` template, it could be required to update the modified template too.

Co-authored-by: christophe dervieux <christophe.dervieux@gmail.com>