Fix Theorem / Proof environments with special markdown content (#1386)
Correctly creates the environment in HTML and LaTeX when no text paragraph first. For example, starting with list or code block

````markdown
::: {#exr-1 .exercise}
1)  Open RStudio.
2)  Write `1 + 9` to console.
:::

::: {.exercise}
```r
library(knitr)
kable(mtcars)
```
Copy and paste the chunk above
:::
````

Special handling is required for LaTeX, and it is done in the Lua filter directly. See about `\leavevmode` in amsthm doc (http://www.ams.org/arc/tex/amscls/amsthdoc.pdf) - it is not the only way of doing it, but it seems to be required to correctly format the result in LaTeX.