Relax `index.Rmd` constraint and correctly support edge case (#1351)
version 0.27 assumes index.Rmd was used in all projects (#1332). It was in fact a recommandation and a requirement only in some scenario. Workflow has now been improved to support `index.Rmd` case-insensitive and other potential main file when running `render_book("main.Rmd")`, as it seems some user are using this workflow (https://community.rstudio.com/t/render-book-error-stop-if-not-exists-and-some-files-were-not-found/139927)