Fix issue with cross referencing figures and tables in epub (#1426)
It seems we were correctly resolving the references, and adding a link to it, but no anchor id anywhere as we do for HTML

This changes insert an anchor id inside an empty span in the caption part. This is a quick solution as the id should be probably added to the figure or table itself.