Tweaking fuse.js configuration to improve search results (#1319)
`ignoreFieldNorm: true` configuration allows to ignore the length of the search fields in the scoring. 

Currently with Fuse.js, the small fields have higher priority in search. This leads to low scoring for longer one, and the search results gets ignored in the output because of score filtering. 

For books, with the help of @jtbayly (thank you), we believe that ignoring the length of search fields make sense as chapters can be long  and then ignored by default.